backend/requirements.txt
django==4.2
djangorestframework==3.14
psycopg2-binary==2.9
python-dotenv==1.0
django-phonenumber-field==7.0
phonenumbers
